+++ Sources/KeyboardKit/Actions/KeyboardAction+Button.swift
@@ -31,7 +31,7 @@ public extension KeyboardAction {
         case .option: return .keyboardOption
         case .primary(let type): return type.standardButtonImage(for: context.locale)
         case .settings: return .keyboardSettings
-        case .shift(let currentCasing): return currentCasing.standardButtonImage
+        case .shift(let currentCasing): return .keyboardShift(currentCasing)
         case .systemImage(_, let imageName, _): return Image(systemName: imageName)
         case .systemSettings: return .keyboardSettings
         case .tab: return .keyboardTab