+++ Sources/KeyboardKit/_Deprecated/InterfaceOrientationResolver.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+#if canImport(UIKit)
+import UIKit
+#endif
+
+@available(*, deprecated, message: "This protocol is no longer used")
+protocol InterfaceOrientationResolver {
+
+    /**
+     Get the current interface orientation
+     */
+    var interfaceOrientation: InterfaceOrientation { get }
+}