+++ Sources/KeyboardKit/_Deprecated/KeyboardCase+Deprecated.swift
@@ -1,18 +1,8 @@
-//
-//  KeyboardCase+Button.swift
-//  KeyboardKit
-//
-//  Created by Daniel Saidi on 2020-07-01.
-//  Copyright © 2020-2023 Daniel Saidi. All rights reserved.
-//
-
 import SwiftUI
 
 public extension KeyboardCase {
     
-    /**
-     The casing's standard button image.
-     */
+    @available(*, deprecated, message: "Use Image.keyboardShift(:) instead")
     var standardButtonImage: Image {
         switch self {
         case .auto: return .keyboardShiftLowercased